[US_UT] Ingest supervision contacts (Recidiviz/recidiviz-data#42025)
## Description of the change

This ingests supervision contacts in UT. We need this to check a
criteria for time cuts that requires a person's address is verified.

That can be confirmed in a few ways using this ingested data:
1) Most specifically: There are contacts with contact type raw text
`ADDRESS VERIFICATION`, contact status `COMPLETED`, and an address
listed in the contact metadata.
2) More generally: There are contacts of any type with contact status
`COMPLETED` with an address in the contact metadata.
3) Most generally: There are contacts with contact type raw text
`ADDRESS VERIFICATION` and contact status `COMPLETED` for a given person
during the time period in question, even if there is no specific address
in the metadata.

[sandbox
pipeline](https://console.cloud.google.com/dataflow/jobs/us-east1/2025-05-05_16_45_55-11863346120893525550;graphView=0?project=recidiviz-staging&inv=1&invt=Abwntw&pageState=(%22dfTime%22:(%22l%22:%22dfJobMaxTime%22)))

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 8a63d98c6ad7a6902f1df8d1a73b200a951afcf8