[US_IA] Refactor supervision violations (Recidiviz/recidiviz-data#41723)
## Description of the change

This PR refactors supervision violations ingest in IA in the following
ways:
- restructures the relationship between incidents, reports, and hearing
decisions
- for every violation incident, we ingest a `StateSupervisionViolation`
object and a `StateSupervisionViolationResponse` object with response
type = "CITATION"
- for every violation report, we ingest a
`StateSupervisionViolationResponse` object with response type =
"VIOLATION_REPORT"
- for every violation hearing decision, we ingest a
`StateSupervisionViolationResponse` object with response type =
"PERMANENT_DECISION"
- **NOTE**: since the relationship between incidents, reports, and
hearings in IA's data is many to many to many, but the relationship
between violations and violation responses in our schema is one to many,
we ingest each root `StateSupervisionViolation` as a unique
incident-report combination and then build relevant violation responses
off of that.
- adds in a view called `supervision_violation_incidents` that ingests
violation incidents without an associated parole or probation violation
report. Like above, for every incident we ingest via this view, we
ingest a `StateSupervisionViolation` object and a
`StateSupervisionViolationResponse` object with response type =
"CITATION"
- adds in parole decisions now that we have the data for parole hearings
data
- removes the fields in `violation_metadata` here since I realized the
information is actually only relevant to the
`StateSupervisionViolationResponse` and not the
`StateSupervisionViolation`

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 88b5b47dd086dbd0176aba86d64ad371b41fb9a4