[US_UT] Add LS/RNR drug & alcohol questions to assessment metadata (Recidiviz/recidiviz-data#42122)
## Description of the change

This adds a number of JSON values to the `state_assessment` metadata
field so that the questions, answers, and comments related to the
"Alcohol/Drug Problem" section of the LS/RNR assessment can be attached
to the assessment result itself in our schema. The section is 1.6
[here](https://drive.google.com/file/d/1PVPnFj4tfXOlkEyhgDrmh--9qagCkq1P/view).

For every other type of assessment, and for LS/RNR assessments entered
into the system before these fields were required to be filled out,
these fields will all be blank in the ingest view results. The metadata
field will have the same JSON keys, but all of the values will be blank.

[Sandbox ingest
pipeline](https://console.cloud.google.com/dataflow/jobs/us-east1/2025-05-07_12_02_20-9261796016917649894;graphView=0?project=recidiviz-staging&inv=1&invt=Abwx4g&pageState=(%22dfTime%22:(%22l%22:%22dfJobMaxTime%22)))
## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: f34538ba0e434d10c9b4c943a1c607d9c391262b