Trecidiviz/ia fx1 sp (Recidiviz/recidiviz-data#42058)
## Description of the change

> We need to add those with pre-trial supervision status (`Pretrial
Release Without Supervision` or `Pretrial Release With Supervision)` to
supervision periods so that analysts can pull pre-trial status into
Early Discharge decision. Specifically we are:
* Adding pre-trial statuses to view filter
* Updating admission and termination reasons enums in the .yaml

Noting that for this view the following time spans are not expected to
have overlapping elements after 2003 migration.
* supervision_level_spans: Except they are and sometimes the overlapping
SupervisionStatusInformationId have conflicting SupervisionLevel
* supervision_modifier_spans: 2024-5 each have 100's of unique
supervisions with overlapping modifiers on the same day.
* supervision_officer_spans
* supervision_location_spans

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 201ebd0a9c5afcc64b5a0ccbb54aae9b0abcfaad