[TES] Refactor TaskCriteriaGroupBigQueryViewBuilder into smaller components (Recidiviz/recidiviz-data#42146)
## Description of the change

Before this change, our view builder class structure looked like this:
<img width="1295" alt="image"
src="https://github.com/user-attachments/assets/b4eff177-e679-44d5-b3f3-d3139e3c9927"
/>

The current structure is:
a) Confusing - There are classes that are called `*BigQueryViewBuilder`
that aren't actual subclasses of `BigQueryViewBuilder`
b) Lossy - if a `OrCriteriaGroup` is converted to an actual view
builder, we have no way to understand the original criteria that went
into the group.

This PR takes a first step towards refactoring the classes so that we
end up with a structure like this:

<img width="1282" alt="image"
src="https://github.com/user-attachments/assets/0bcab98e-c035-4ef1-8c72-2e305ce3a218"
/>

In this new world, every sub-criteria will be an actual
`BigQueryViewBuilder`, which will simplify logic in a number of places,
and also the new view builder classes like
`StateSpecificTaskCriteriaGroupBigQueryViewBuilder` will have a place to
store the original list of criteria used to generate that view.

This PR:
* Creates a helper class, `TaskSubCriteriaGroup` that houses logic just
related to generic understanding of a collection of sub-criteria.
* Moves all group criteria query building logic to a new class

When rebased on the latest alpha, this refactor produced zero view
changes when running a sandbox in auto mode.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Pre-work for Recidiviz/recidiviz-data#41711

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: e49ac4dd16c46d9017dfc5c30b3100beab5bb710