[US_AR] Add deduping step for state person alias (Recidiviz/recidiviz-data#42170)
## Description of the change

Ingest pipelines are failing because someone has multiple instances of
the same alias with the same alias type (INTERNAL_UNKNOWN). This can
happen because while a given alias can only show up once in the raw
table for each possible alias type code, multiple of these alias type
codes are mapped to INTERNAL_UNKNOWN. This change adds a deduping step
to ensure that if there are multiple records with the same alias that
all would otherwise all get mapped to alias type INTERNAL_UNKNOWN, only
one of them is kept.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 7a1a354ee6a03c31323c805a91013e7abf7ad29e