[US_IA][Workflows] Create criterion to exclude clients on pre-trial investigation  (Recidiviz/recidiviz-data#42135)
## Description of the change

Per early discharge policy, people with pending charges should not be
eligible. This includes people with pretrial investigation statuses (as
verified by TTs). Note - the `INVESTIGATION` supervision type is not yet
ingested for Iowa, but Tare is working on it! In the meantime I checked
the results using an ingest sandbox
(`recidiviz-staging.tsur_ia_us_ia_normalized_state.state_supervision_period`)
and found that 75 currently eligible people are on pretrial status. I
also checked that the 2 examples that we discussed in a recent TT call
of people being on pretrial status are included in this criterion.

Totally unrelated but I also removed a TODO related to missing
supervision levels in Iowa - I took another look and I think this has
been resolved with the ingesting of new data.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 501763b2cc9d1fa664d94ddf7d29c4aa7f434426