[US_TX] Tweaks TC phases logic (Recidiviz/recidiviz-data#42239)
## Description of the change

This PR tweaks the TC phase logic. We no longer look at past referrals,
nor do we consider the phases in the team meetings.

The new flow is as follows:
1. Check whether client is in an in-patient facility, if so then "Phase
1"
2. Check whether client has a peer support referral and no other active
program referrals, if so then "Phase 3"
3. Else, "Phase 2"

Testing was done in this google sheet:
https://docs.google.com/spreadsheets/d/1UvfNcrgFUVyvwYdy-0MQfytujTJxkcRYNTFDHNv2nwY/edit?gid=0#gid=0

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 97a87eab63765b6d546b52ec73da7d245675ebcb