add new product_roster_archive view with enabled FVs (Recidiviz/recidiviz-data#42142)
## Description of the change

As part of https://github.com/Recidiviz/recidiviz-data/issues/39162, we
want to be able to tell if a user had an FV enabled on a given day. This
PR creates a more traditional archive view for `product_roster`
(treating the `export_archive` one as raw input to it and having this be
the one to be used everywhere), filters out null email addresses which
are pretty useless, and adds a new column `enabled_fvs` to determine
which feature variants were actually enabled for a user on a given day.

BigQuery's JSON support doesn't have as much functionality as
postgres's, so a lot of what's going on with the regex parsing and CTEs
is trying to imitate the `default_feature_variants ||
override_feature_variants` JSON concatenation that happens
[here](https://github.com/Recidiviz/recidiviz-data/blob/main/recidiviz/auth/auth_users_endpoint.py#L168-L173).
It's not a perfect system because it splits on commas, which means if we
add more stuff to the FV object type it'll break, but we haven't done
that basically ever, and there doesn't seem to be a way to do better
outside of using user-defined functions (which I remember trying to do
ages ago and it didn't work- I think it's not compatible with views
maybe?)

There wasn't a super obvious single reviewer for this, so putting in a
grab bag of people who have at least some context! Let me know if you
have questions!

Loaded to sandbox at `dana_roster_archive` (did reference_views only)

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

https://github.com/Recidiviz/recidiviz-data/issues/40650

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 8359db4ea100ad4bee0a07f0cea504bf8e0ad07c