[raw data] adds `max_num_unparseable_bytes` raw file config property, `LimitedErrorReplacementHandler` class (Recidiviz/recidiviz-data#42278)
## Description of the change

this pr does two things:
1. adds a new yaml property `max_num_unparseable_bytes` to raw file
configs
2. adds the scaffolding for cautiously ignoring unparseable bytes with
`LimitedErrorReplacementHandler`. the idea here is that we can just
register this for each file we need to parse that has
`max_num_unparseable_bytes` specified for each chunk and then use our
normal `TextIOWrapper` in `DirectIngestRawFilePreImportNormalizer`.
then, we can coalesce all the unparseable bytes for each chunk at the
end to figure out how many were in total (if any individual chunks is
above the limit, we can throw right then)

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

prework for Recidiviz/recidiviz-data#42277

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 20f8a512aa2632c72d87b166d80689b4a42410f0