[TES] Convert TaskCriteriaGroupBigQueryViewBuilder into real view builder classes (Recidiviz/recidiviz-data#42204)
## Description of the change

Similar to Recidiviz/recidiviz-data#42187, but for criteria groups. Deletes
`TaskCriteriaGroupBigQueryViewBuilder` in favor of
`StateSpecificTaskCriteriaGroupBigQueryViewBuilder` and
`StateAgnosticTaskCriteriaGroupBigQueryViewBuilder` classes.

At this point, all complex criteria extend from
`StateSpecificTaskCriteriaBigQueryViewBuilder` and
`StateAgnosticTaskCriteriaBigQueryViewBuilder`, which makes a lot of the
mypy typing simpler.

I tested and this produced no sandbox changes.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#41711

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: fcc0c7fe1409d131f9e33a2300fa1ab9d1d2abbe