[Insights] add is_surfaced_category_type check to outlier archive sessions (Recidiviz/recidiviz-data#42345)
## Description of the change

In a different PR I'm working on, I noticed some single-day spans in
`insights_supervision_officer_outlier_status_archive_sessions`, which
maybe itself is not a problem but was at least surprising to me. I
noticed that some of these seemed to potentially stem from cases where
we have multiple category types calculated for a metric (and so
appearing in the archive), but we aren't doing any deduping on that in
the archive sessions view, and so it may have been messing with our
numbers a bit. This excludes metrics that weren't the primary category
type displayed in the tool on that day, since I think we don't actually
care about those in any views that would be downstream of this.

This brings the # of single-day spans in this view down to 47 from 6278.
Loaded to sandbox at `dana_outlier_status_archive_category`

Also after writing ^, I discovered I had already written that I wanted
to do this in
https://github.com/Recidiviz/recidiviz-data/issues/40790#issuecomment-2798196623,
but had forgotten 🤦‍♀️

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#40790

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 46c2909b7b7dbdc237fc4161b82f7a698ed892b2