[US_AZ] Add data source for manlit criteria (Recidiviz/recidiviz-data#42184)
## Description of the change

This PR adds a data source for checking if a person has met ADCRR's
mandatory literacy requirement.

Rather than using actual TABE testing data, we were directed to use the
DOC Priority Report for accessing the variable `MEET_STANDARD` to
understand if a person has met ADCRR's mandatory literacy requirement.
The table itself is used for programming enrollment prioritization, but
the research team told me this field always relates to the mandatory
literacy requirement.

I am going to keep pushing for TABE data for the sake of completeness,
but I think this is a good proxy to solve user-reported errors in the
meantime.

I loaded the change to
`recidiviz-staging.elise_task_eligibility_spans_us_az.all_tasks_materialized`
and observed that it makes no difference to eligibility. I also
confirmed that the client who was flagged in the original user report
has since been released on TPR, so he wouldn't show up as newly
eligible, but I also confirmed that this change does make him appear as
having completed the literacy requirement.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: c32e40d498fe05c012c5af0086128c0f70aa3d31